requirements.txt
absl-py==0.11.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.21
fastapi==0.63.0
google-auth==1.26.1
google-auth-oauthlib==0.4.2
grpcio==1.35.0
h11==0.12.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.2
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
Werkzeug==1.0.1
